# server/openapi/parallel-contracts.yaml
openapi: 3.0.0
info:
  title: Reserves v1.1 Parallel Implementation API
  version: 1.1.0
  description: Contract-first API specification for parallel development
servers:
  - url: http://localhost:5000/api
    description: Development server
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: updog.session
      description: Browser HttpOnly JWT cookie; unsafe requests also require X-CSRF-Token.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Signed jti-bound CSRF token for cookie-authenticated unsafe requests.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Machine/service JWT transport; browser code must not use this scheme.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Unique key for idempotent operations
    XOrgId:
      name: X-Org-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Organization context for RLS
...
  schemas:
...
    ApprovalResponse:
      type: object
      properties:
        approvalId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        requiredSignatures:
          type: integer
        currentSignatures:
          type: integer
        expiresAt:
          type: string
          format: date-time
paths:
  /auth/logout:
    post:
      summary: End the current authenticated session
      description: Cookie-authenticated requests require both updog.session and X-CSRF-Token; machine Bearer requests are CSRF-exempt.
      security:
        - CookieAuth: []
          CsrfToken: []
        - BearerAuth: []
      responses:
        '204':
          description: Session cookies cleared and token revocation recorded
        '401':
          description: Missing, invalid, or ambiguous credentials
        '403':
          description: Cookie-authenticated request failed CSRF validation
        '503':
          description: Cookies cleared but durable token revocation failed

  # ==========================================================
  # CONCURRENCY ENDPOINTS (Step 2)
  # ==========================================================
...
  /reserves/calculate:
    post:
      summary: Calculate reserves with versioning and protection
      security:
        - CookieAuth: []
          CsrfToken: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XOrgId'
        - name: version
          in: query
          schema:
            type: string
            pattern: '^\d+\.\d+\.\d+$'
          description: Specific version to use
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Calculation completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                  metadata:
                    type: object
                    properties:
                      version:
                        type: string
                      wasmSha256:
                        type: string
                      executionTimeMs:
                        type: integer
                      deterministic:
                        type: boolean
                      approvalId:
                        type: string
        '409':
          description: Fund busy or idempotency conflict
        '428':
          description: Approval required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalResponse'
...